Relocate which_docker to environment module
This function is more related to inspecting the host state than the
docker-daemon specificaly.  Moving it also supports easier unittesting
w/o many unrelated mock()'s.  It's still imported by name into the
docker_daemon module, so no API breakage is possible.
